Ensure project and team viewsets are protected by ownership
